Customized boot.sh from rails/rails

+++ .devcontainer/boot.sh
@@ -1,8 +1,16 @@
-bundle install --path=vendor/bundle
-yarn install
+#!/bin/bash
+
+[[ -f "Gemfile" ]] && bundle install --path=vendor/bundle
+[[ -f "yarn.lock" ]] && yarn install
 
 sudo service postgresql start
 sudo service redis-server start
 
 # Create PostgreSQL users and databases
 sudo su postgres -c "createuser --superuser vscode"
+
+# Setup the rails DB if in a Rails app.
+[[ -x "bin/rails" ]] && bin/rails db:setup
+[[ -x "bin/rails" ]] && bin/rails stats
+
+exit 0
